Moves all tests to single file
