+++ SnapshotTests/ReferenceImages_64/UnitTests.DemoComponentsTests/testCheckboxDemoView@3x.png
