+++ SnapshotTests/ReferenceImages_64/UnitTests.DemoComponentsTests/testConsentTransparencyInfoDemoView@3x.png
