+++ SnapshotTests/ReferenceImages_64/UnitTests.DemoComponentsTests/testEmptyViewDemoView@3x.png
