+++ SnapshotTests/ReferenceImages_64/UnitTests.DemoComponentsTests/testFrontpageViewDemoView@3x.png
