+++ SnapshotTests/ReferenceImages_64/UnitTests.DemoComponentsTests/testLabelDemoView@3x.png
