+++ SnapshotTests/ReferenceImages_64/UnitTests.DemoComponentsTests/testRadioButtonDemoView@3x.png
