+++ SnapshotTests/ReferenceImages_64/UnitTests.DemoComponentsTests/testRibbonDemoView@3x.png
