+++ SnapshotTests/ReferenceImages_64/UnitTests.DemoComponentsTests/testRoundedImageViewDemoView@3x.png
