+++ SnapshotTests/ReferenceImages_64/UnitTests.DemoComponentsTests/testSwitchViewDemoView@3x.png
