--- Sources/Components/Consent/ConsentTransparency/ConsentTransparencyInfoView/Demo/ConsentTransparencyInfoDemoViewTest.swift
@@ -1,21 +1,0 @@
-//
-//  Copyright © 2018 FINN AS. All rights reserved.
-//
-
-import FBSnapshotTestCase
-import FinniversKit
-import Demo
-
-class ConsentTransparencyInfoDemoViewTest: FBSnapshotTestCase {
-
-    override func setUp() {
-        super.setUp()
-        recordMode = false
-    }
-
-    func testConsentTransparencyInfoDemoView() {
-        let controller = ViewController<ConsentTransparencyInfoDemoView>()
-        FBSnapshotVerifyView(controller.view)
-    }
-
-}
