--- Sources/Components/SelectionBoxes/Demo/CheckboxDemoViewTest.swift
@@ -1,20 +1,0 @@
-//
-//  Copyright © 2018 FINN AS. All rights reserved.
-//
-
-import FBSnapshotTestCase
-import FinniversKit
-import Demo
-
-class CheckboxDemoViewTest: FBSnapshotTestCase {
-
-    override func setUp() {
-        super.setUp()
-        recordMode = false
-    }
-
-    func testCheckboxDemoView() {
-        let controller = ViewController<CheckboxDemoView>()
-        FBSnapshotVerifyView(controller.view)
-    }
-}
