--- Sources/Components/Switch/Demo/SwitchViewDemoViewTest.swift
@@ -1,21 +1,0 @@
-//
-//  Copyright © 2018 FINN AS. All rights reserved.
-//
-
-import FBSnapshotTestCase
-import FinniversKit
-import Demo
-
-class SwitchViewDemoViewTest: FBSnapshotTestCase {
-
-    override func setUp() {
-        super.setUp()
-        recordMode = false
-    }
-
-    func testSwitchViewDemoView() {
-        let controller = ViewController<SwitchViewDemoView>()
-        FBSnapshotVerifyView(controller.view)
-    }
-
-}
